change in validation in backend

+++ app/Http/Controllers/ChangePasswordController.php
@@ -2,15 +2,14 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
 use App\Models\ResetPassword;
-use App\Http\Requests\ChangePasswordRequest;
 use Symfony\Component\HttpFoundation\Response;
 use App\Models\User;
+use Symfony\Component\HttpFoundation\Request;
 
 class ChangePasswordController extends Controller
 {
-    public function process(ChangePasswordRequest $request)
+    public function process(Request $request)
     {
         return ($this->getPasswordResetTableRow($request) && !empty($request->verify)) ? $this->changePassword($request) : $this->tokenNoFoundResponse();
     }
@@ -29,7 +28,11 @@ class ChangePasswordController extends Controller
 
     private function changePassword($request)
     {
-        $user = User::whereEmail($request->email)->first();
+        if(!$user = User::whereEmail($request->email)->first()){
+            return response()->json([
+                'error' => 'Email invalid'
+            ], Response::HTTP_UNPROCESSABLE_ENTITY);
+        }
         $user->update(['password' => $request->password]);
         $this->getPasswordResetTableRow($request)->delete();
         return response()->json([
